requirements.txt
sentence-transformers==2.7.0
psycopg[binary]>=3.1.0
pgvector>=0.3.0
pdfplumber==0.11.0
pymupdf==1.24.0
pytesseract==0.3.10
Pillow==10.3.0
tqdm==4.66.4
numpy==1.26.4
torch==2.3.0
python-dotenv==1.0.1
rich==13.7.1
